Support Ecto v3.4.2 and rewrite ExAudit.Repo to extend Ecto.Repo
This is a breaking change to support ecto and ecto_sql ~> 3.4 and at the
same time to actually use ExAudit.Repo with Ecto.Repo.

The Ecto.Repo @callback implementations are overriden to add version tracking if the
given struct or changeset is in the list of :tracked_schemas specified in the config
or call the original Ecto.Repo implementation otherwise.